requirements.txt
argon2-cffi==20.1.0
cffi==1.14.2
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycparser==2.20
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
